// src/components/Layout/types.ts
export interface ILayoutProps {
  gapDefault?: string;
  marginDefault?: string | number;
  gapMedium?: string;
  marginMedium?: string | number;
  gapSmall?: string;
  marginSmall?: string | number;
}

export interface IRowProps {
  xss: number;
  xse: number;
  sms: number;
  sme: number;
  mds: number;
  mde: number;
  lgs: number;
  lge: number;
  xls: number;
  xle: number;
  xxls: number;
  xxle: number;
}